% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sumstats.R
\name{baba}
\alias{baba}
\title{Calculate the baba portion of the D-statistic}
\usage{
baba(p1, p2, p3)
}
\arguments{
\item{p1}{is a numeric vector with the allele frequencies for population 1.
Each entry of the vector should be a different site.}

\item{p2}{is a numeric vector with the allele frequencies for population 2.
Each entry of the vector should be a different site.}

\item{p3}{is a numeric vector with the allele frequencies for population 3.
Each entry of the vector should be a different site.}
}
\value{
a numeric vector with one value per site of the locus.
}
\description{
Computes the value for the ‘BABA’ allelic pattern.
}
\keyword{internal}
