% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_import.R
\name{createHeader}
\alias{createHeader}
\title{Create a header for a _rc file of popoolation2}
\usage{
createHeader(nPops)
}
\arguments{
\item{nPops}{is an integer specifying how many different populations exist in
the _rc file.}
}
\value{
a character vector with the column names for a _rc popoolation2 file.
}
\description{
Creates a header for files in the _rc format of the \code{popoolation2}
software. This header can be applied to a matrix as column names.
}
\details{
Please note that the first 9 columns are a default output of the
\code{popoolation2} software and thus this functions maintains the same
names.
}
\examples{
createHeader(nPops = 10)

}
