% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sumstats.R
\name{D.stat}
\alias{D.stat}
\title{calculate D-statistic}
\usage{
D.stat(ABBA, BABA)
}
\arguments{
\item{ABBA}{is a numeric vector with the values for the ‘ABBA’ allelic
pattern.}

\item{BABA}{is a numeric vector with the values for the ‘BABA’ allelic
pattern.}
}
\value{
a numeric value representing the D-statistic value for a particular
conformation of the populations.
}
\description{
Computes the value of the D-statistic given the values of the two particular
allelic patterns.
}
\keyword{internal}
