% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ABC.R
\name{euclidean}
\alias{euclidean}
\title{Compute euclidean distance}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{is an integer or a numeric vector.}

\item{b}{is another integer or numeric vector.}
}
\value{
a numeric value representing the euclidean distance between the two
inputs.
}
\description{
Computes the euclidean distance between two integers or numeric vectors.
}
\keyword{internal}
