% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sumstats.R
\name{Expected_Het}
\alias{Expected_Het}
\title{Compute expected heterozygosity within a population}
\usage{
Expected_Het(Pop_Pi)
}
\arguments{
\item{Pop_Pi}{is a matrix or list of allele frequencies. When dealing with a
single locus, this input is a matrix and when dealing with multiple loci it
is a list. Each entry of that list is a matrix representing a different
locus. Each row of that matrix should correspond to a different population
and each column to a different SNP.}
}
\value{
if the input is a single matrix, the output will be a matrix where
each row represents a different population and each column is the expected
heterozygosity of a population at that site. If the input is a list, the
output will also be a list, with each entry corresponding to a different
locus. Each of those entries will be a matrix with different populations in
different rows and the expected heterozygosity of different sites at
different columns.
}
\description{
This functions calculates the value of the expected heterozygosity for each
SNP.
}
\keyword{internal}
