% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myparams}
\alias{myparams}
\title{Matrix of simulated parameter values}
\format{
a matrix with 5000 rows and 8 columns:

\describe{

\item{N1}{relative size of the first population. This population
corresponds to the C ecotype.}

\item{N2}{relative size of the second population. This population
corresponds to the W ecotype.}

\item{Split}{time, in 4Nref scale, of the split event that creates the two
populations.}

\item{PoolError}{error associated with DNA pooling.}

\item{SeqError}{error associated with DNA sequencing.}

\item{mCW}{migration rate between the two divergent ecotypes This
is the migration rate from ecotype C to ecotype W.}

\item{mWC}{migration rate between the two divergent ecotypes This
is the migration rate from ecotype W to ecotype C.}

\item{pM}{proportion of the genome with no barriers against gene flow. This
is the proportion of simulated loci where migration occurs in both
directions between the divergent ecotypes.} }
}
\source{
simulations performed
}
\usage{
myparams
}
\description{
This data set contains a matrix of simulated parameter values.
These parameter values were sampled from prior distributions and used to
perform simulations under a isolation with migration model with two
populations. Each row of this matrix corresponds to a different simulation.
}
\keyword{datasets}
