% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ABC.R
\name{normalise}
\alias{normalise}
\title{Normalize data - adjust values measured on different scales}
\usage{
normalise(x, y)
}
\arguments{
\item{x}{is a numeric vector.}

\item{y}{is a numeric vector.}
}
\value{
a numeric vector with the same length as \code{x}.
}
\description{
Adjusts values that are measured over different scales.
}
\details{
This function is used to scale the summary statistics and the target for ABC
inference. This scaling ensures that the different summary statistics are all
in the same scale when performing model selection or parameter inference.
}
\keyword{internal}
