% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ABC.R
\name{removeVar}
\alias{removeVar}
\title{Remove columns with zero variance}
\usage{
removeVar(observed, sumstats)
}
\arguments{
\item{observed}{is a matrix of observed summary statistics. Each column
should be a different summary statistic.}

\item{sumstats}{is a matrix of simulated summary statistics. Each column
should be a different summary statistic and each row a different
simulation.}
}
\value{
a list with two named entries. One entry contains the matrix of
observed summary statistics and the other the simulated summary statistics.
}
\description{
Removes summary statistics with zero variance.
}
\details{
Checks the variance of the summary statistics in the observed data and
removes summary statistics with zero variance. Those summary statistics are
removed from both the matrix of observed values and the matrix of simulated
values.
}
\keyword{internal}
