% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumstats}
\alias{sumstats}
\title{Matrix of summary statistics computed from simulated data}
\format{
a matrix with 10000 rows and 14 columns:

\describe{

\item{Sf}{fraction of sites fixed between populations.}

\item{Sx1}{fraction of exclusive sites for the first population.}

\item{Sx2}{fraction of exclusive sites for the second population.}

\item{SS}{fraction of sites shared between the two populations.}

\item{Mean_Het1}{mean expected heterozygosity of the first population.}

\item{Mean_Het2}{mean expected heterozygosity of the second population.}

\item{SD_Het1}{standard deviation across loci of the mean expected
heterozygosity of the first population.}

\item{SD_Het2}{standard deviation across loci of the mean expected
heterozygosity of the second population.}

\item{Mean_HetBet}{mean heterozygosity between the two populations.}

\item{SD_HetBet}{standard deviation across loci of the mean heterozygosity
between the two populations.}

\item{Mean_FST}{mean pairwise FST between the two populations.}

\item{SD_FST}{standard deviation across loci of the mean pairwise FST
between the two populations.}

\item{FSTQ1}{5\% quantile of the mean pairwise FST distribution.}

\item{FSTQ2}{95\% quantile of the mean pairwise FST distribution.} }
}
\source{
simulations performed
}
\usage{
sumstats
}
\description{
This data set contains a set of 14 summary statistics computed
from data simulated under an isolation with migration model of two
populations.
}
\keyword{datasets}
