% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ABC.R
\name{tranf}
\alias{tranf}
\title{Apply a transformation to the parameters}
\usage{
tranf(x, min, max)
}
\arguments{
\item{x}{is the parameter vector (long vector of numbers from the
simulations). These are the values to be transformed.}

\item{min}{is the minimum value of the prior for this parameter.}

\item{max}{is the maximum value of the prior for this parameter.}
}
\value{
a numeric vector with the same length as \code{x} with the parameter
values transformed.
}
\description{
This function applies a transformation to the parameter values.
}
\details{
The transformation should be applied before parameter estimation using an
Approximate Bayesian Computation framework to ensure that the estimates do
not fall outside the boundaries set by the prior distribution.
}
\keyword{internal}
