% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_simulations.R
\name{runSCRM}
\alias{runSCRM}
\title{Run scrm and obtain genotypes}
\usage{
runSCRM(commands, nDip, nPops, model)
}
\arguments{
\item{commands}{A character string containing the commands for the scrm
package. This string can be created using the \code{cmd2pops}, the \code{cmdSingle}
or the \code{cmdParallel} functions.}

\item{nDip}{An integer representing the total number of diploid individuals
to simulate. Note that scrm actually simulates haplotypes, so the number of
simulated haplotypes is double of this.}

\item{nPops}{An integer that informs of how many populations exist on the
model you are trying to run.}

\item{model}{Either "2pops", "Single" or "Parallel" indicating which model
should be simulated.}
}
\value{
a list with the simulated genotypes. Each entry is a different locus
and, for each locus, different rows represent different individuals and
each column is a different site.
}
\description{
This function will run the scrm package, according to the command line
supplied as input. It will also combine haplotypes into genotypes and
re-organize the output if the simulations were performed under a single
origin scenario. This is to ensure that the output of the four-population
models will always follow the same order: the two divergent ecotypes in the
first location, followed by the two divergent ecotypes in the second
location.
}
